Raster fix (iss124)

+++ src/drawers/gcodedrawer.h
@@ -115,7 +115,7 @@ private:
     int getSegmentType(LineSegment *segment);
     QVector3D getSegmentColorVector(LineSegment *segment);
     QColor getSegmentColor(LineSegment *segment);
-    void setImagePixelColor(QImage &image, int x, int y, QRgb color) const;
+    void setImagePixelColor(QImage &image, double x, double y, QRgb color) const;
 };
 
 #endif // GCODEDRAWER_H
